+++ lib/windows/windows_tznames.zig
@@ -3,7 +3,7 @@
 //
 // --- Do not edit ---
 //
-// latest referesh: 2025-01-16T18:56:31+00:00
+// latest referesh: 2025-03-23T08:05:29+00:00
 // windows_names are sorted alphabetically so we can do binary search (later)
 pub const windows_names = [_][]const u8{
     "AUS Central Standard Time",